Add support for REG_PCIE registers via debugfs
Updated soc15_parse.sh to read smn registers.  Added pcie read/write
support to the lowlevel mmio routines.  Added support for the use
of the word 'pci' and 'pcie' in update scripts and added support for
update commands on the command line (instead of having to use a separate file).

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>