Add support for ZFB and fix PTE-FURTHER
This commit also cleans up a lot of the logic to make maintenance
easier.

Supprort for PTE-FURTHER was broken in a recent commit.  We now
read the BLOCK_FRAGMENT_SIZE from PDE0 correctly (instead of BASE) which
then changes the PAGE_TABLE_BLOCK_SIZE to accomodate resized PTBs.

A new test case covers PDE/PTE's with S=0 (which previously had failed)

A new test case for PTE-FURTHER (which recently failed)

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>