Introduction of --ring-stream command
This patch introduces a new ring decoder for the Linux
application that uses the stream decoder for pm4 and
sdma rings.  It is meant to be a drop in replacement
for the --ring command.

The purpose of this patch is to obsolete the existing
older FSM decoders to reduce the size of the code base
and maintenance effort required long term.

TODO: MQD decoders

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>