Upgrade MMIO register support
The kernel supports a new debugfs file (amdgpu_regs2) which
uses an IOCTL to set state and then a full 32-bit address
space for MMIO.

This patch adds support for that while retaining support
for the older path.  Eventually, the old path will be removed.

This patch also fixes a few other things

- Support for 64-bit "--write-bit" registers
- Drop the grant()/release() callbacks
- Drop a lot of inline lseek/read/write calls to direct to
  the reg_funcs callbacks
- Simplify scan/set_reg/set_bit code to use callbacks for both pci/debugfs

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>