Parse SDMA with more up to date packet definition
This brings COPY LINEAR, COPY BROADCAST, COPY TILED, COPY SOA, COPY
LINEAR_SUB_WINDOW packets up to date with gfx10. It looks like the
original definitions used were from tonga and do not include changes
that were made for Vega and Navi. Some of the packet sizes changed so
without this patch the parser fails to parse buffers with these packets
in them.

This also includes a fix for parsing WRITE LINEAR packet. The number of
DWORDs in the packet to write is given by the COUNT field + 1.

Change-Id: Id8c7759977d99421383967a51316eed921d75886
Signed-off-by: Colin Lee <Colin.Lee@amd.com>
Reviewed-by: Tom St Denis <tom.stdenis@amd.com>