Clean up failed ASIC initialization
When files are missing the error messages were too short
to be helpful.  This adds some instruction on the database
as well as fixes a segfault on failed init.

Finally, this also prevents trying to init by PCI DID
if a corresponding named .asic file exists but is damaged
or otherwise not useable.

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>