Add ui for returning size of unhandled SDMA packets
The main thing is that the ui originally used for
umr_sdma_decode_stream_opcodes is now also used for
umr_sdma_decode_stream to include an interface for returning the size of
a packet with a given op code. This is needed because SDMA packets do
not the size of the packet in the packet header, so the size of the
packet is only knowable by the opcode. The new callback allows
externally returning the size of the packet.

Signed-off-by: Colin Lee <Colin.Lee@amd.com>