server, gui: add a way to pass raw data alongside json
Passing raw data as JSON_Array of JSON_numbers can fail when
interpreting the raw bits produces an invalid double.

In this case, parson would drop the value and this would break
decoding randomly.

This commit instead adds a way to pass raw data and adapt
both the server and the gui to take advantage of this feature.

Signed-off-by: Pierre-Eric Pelloux-Prayer <pierre-eric.pelloux-prayer@amd.com>