Make PM4 decoder a little more robust
Now start_opcode() is not called when it's an unhandled opcode.
This is meant to be more consistent with the other decoders such
as the SDMA decoder.

We also reject type 0 packets on anything above AI+ since the
firmware doesn't support it.  Any type 0 packet or packet
with an invalid type field will now be marked as invalid (0xff)
and passed off to a new unhandled_dword() ui callback.

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>