Wave scanning re-write
This patch drops the hardcoded umr_wave_status structure members in favour
of a per GFX version list of wave registers and an API to make access
to various flags and datums more generic.  This is necessary as future
ASICs will diverge in how relevant wave status fields are stored.

As a result of these changes the amdgpu_gprwave debugfs file is also
now mandatory.  It has been around for quite some time so this shouldn't
be a problem.

The --waves command has been enhanced to use this new API as well
as dump data to a temporary file.  This allows it to resume waves (if
halted) before displaying to the user since simply drawing the output to
the console can delay long enough to cause a gpu reset.

Because of timing constraints the GUI has not been updated and is disabled for
now.

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>