Introduction of RLS reading for KFD
This adds the --runlist command which can dump the PM4
runlist attached to a specified node.

This patch also fixes an endless loop on compute bound
machines that had DRI directories with a 'name' file that
included the PCI bus addr but weren't ASIC instance directories.
Now UMR only considers DRI directories that include an AMDGPU debugfs
file (in this case 'amdgpu_regs2').

Eventually the goal is to trap PASIDs and map them to VMIDs and
read queues and other structures stored in memory.

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>