Introduction of Remote UMR (rumr) functionality
This patch introduces the Remote UMR (rumr) client and
example UMR backed server that can be used as a resource
to develop non-umr based servers for non-traditional
environments (such as virtualization).

The client should remain static and it provides most umr
functionality without need for direct access to kernel
facilities such as debugfs/sysfs.  It can also be run
as an unprivileged user.

The provided UMR backed server is an example server
that shows how the various packets are meant to be handled.
It can be used in scenarios where root access isn't available
such as inside a sandbox or docker container.

The patch also fixes up a variety of minor bugs that came out
of the endeavour (such as the -cpc and -sdma commands) as well
made various upgrades (such as adding a ring_read_data() callback and
various hoisted _raw() functions).

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>