reorg rumr code and add serialized asic support to discovery
Now you can export an otherwise discovered ASIC to disk which is
useful for supporting no_kernel mode.  When you use --pci
with no_kernel mode it will try and load one of the DID based
'sasic' files which gives umr not only the IP discovery but the
GCA configuration data.

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>