gui: extend memory panel features
This commits changes the way memory usage was tracked and displayed.
All the memory usage info is now kept, and can be displayed in a new
history view.

The original treemap view is still available.
In the treemap view, BOs are displayed per-fd, which is useful for
apps the open the render node multiple times.

Signed-off-by: Pierre-Eric Pelloux-Prayer <pierre-eric.pelloux-prayer@amd.com>