Fix enumeration on MI hosts with non-GPU DRI devices listed
On some MI hosts UMR would have bound to the incorrect DRI instance
for a given PCI bus address resulting in it failing to IP discovery
and not being terribly useful.

Now we check if IP discovery exists and if not prefer to continue.
Ultimately using the saved entry for pre-IP discovery systems.

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>