Initial user mode queue support for umr (v3)
Introduction of several new commands and behind the scene changes to support user
mode queues.

The --user-queue command is added to help attach to user mode queues either by
client id, command PID, or command name, followed by the queue either by number
or type.

The --print-uq command is added to print out the decoded User Queue data for diagnostics.

The --dump-uq command is added to dump the PM4 packets, by default it reads from
between the RPTR to WPTR.  If you specify the new -O use_full_user_queue option
it will read from 0 to WPTR which is useful if the CP has moved past where the SQ
is blocked.

The --waves (-wa) command was updated to support a 'ring' name of 'uq' which
instructs umr to use the user queue data that --user-queue found.

v2: Made umr_parse_clientid() a lowlevel linux function instead of in src/app and
made it return 0 on success, -1 on error.  Also made it so you can call it multiple
times (by turning off VM translation between calls).  Added a function
umr_read_user_queue_buffer() to read the queues as a ring allowing start > end.
Using this I updated --dump-uq and umr_print_waves() to use it instead.  RUMR
support was upgraded to support user queues (and fix a few subtle bugs that
predated this change).

v3: Added more info to the output of --print-uq and --dump-uq and now use
the process PD address instead of GPU address.

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>