Add KFD Client support to user queues
This patch updates (a lot of things) mainly to add KFD client support
to the --user-queue command.

Firstly, the syntax to specify a client was re-written to be more
human friendly.  Then KFD support was added to read the debugfs
entries specific to KFD.  The MQD decoding was updated to support
gfx9 through gfx12 though only compute queues were tested on
gfx10 through gfx12.  KGD clients are only supported on gfx12
but this hasn't been tested since the initial UQ patch.

The HSA decoder was updated to correctly size the AQL packets
and then also attach shaders found in DISPATCH_KERNEL packets.

The --waves command was updated to print out kernel registers
along with the kernel program text.  For compute queues support
was also added to try and find the DISPATCH packet when the PC
address lies outside the attached kernels text space.

There are a variety of updates to output text in a variety of
commands.

For HMM tasks UMR now supports trying to pread/pwrite() the
client process space which allows accessing pages that aren't bound
to the AMDGPU device.  This happens transparently behind the scenes
when a user queue is attached.

The sphinx documentation was updated to add the new commands and modes
of operation and refresh some of the older content.

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>