Refactor umr_access_vram_ai() step one
There is a plan to tidy up this function to make it easier to work on, maintain,
debug, etc.  The first step was to organize the very many variables into their
own structures, and then consolidate the structures into a larger one so that
when we eventually create new functions to refactor the code we can easily
pass the entire translation state around.

Took the time to also add detailed comments to the variables which helps
both human and LLM users alike.

Signed-off-by: Tom St Denis <tom.stdenis@amd.com>