gui, server: rework BO parsing
amdgpu_gem_info and amdgpu_vm_info sysfs files and lots in common,
so this commit switches to a single function to parsing them.

The main differences between the 2 files are:
* amdgpu_gem_info prints BO handles, while amdgpu_vm_info prints a
  meaningless index.
* gem_info reports a pid (tgid) and command name but it's sometimes incorrect
  (eg: 'systemd-logind' is reported instead of 'gnome-shell'). vm_info
  reports tid and a thread name.

postprocess_gem_info is a new function to correlate the data from
these 2 files, and generate a single output.

The data format is changed to "an applications array, each one containing
a clients array, each one containing a BOs array", so the GUI code is
adapted to these changes.

Signed-off-by: Pierre-Eric Pelloux-Prayer <pierre-eric.pelloux-prayer@amd.com>